Moved auth0 definition to plugins folder
