+++ src/main.ts
@@ -4,7 +4,7 @@ import App from "./App.vue";
 import router from "./router";
 import Store from "./store";
 import mediaPlayer from "./plugins/mediaPlayer";
-import auth0 from "./auth0";
+import auth0 from "./plugins/auth0";
 import bmmApi from "./plugins/bmm-api";
 import i18n from "./plugins/i18n";
